+++ public/_redirects
@@ -1,2 +1,1 @@
-/     /list         302!
 /*    /index.html   200
