+++ public/_redirects.extra-1
@@ -1,0 +1,2 @@
+/     /discover         302!
+/*    /index.html       200
