+++ public/serve.extra-1.json
@@ -1,0 +1,9 @@
+{
+  "redirects": [
+    {
+      "source": "/",
+      "destination": "/discover",
+      "type": 302
+    }
+  ]
+}
