+++ scripts/build-variants.js
@@ -39,7 +39,7 @@ function go() {
     baseRoute = baseRoute.endsWith('/') ? baseRoute : `${baseRoute}/`
     baseRoute = baseRoute.startsWith('/') ? baseRoute : `/${baseRoute}`
     return `
-${baseRoute}        ${baseRoute}list              302!
+${baseRoute}        ${baseRoute}discover              302!
 ${baseRoute}*       ${baseRoute}index.html        200
     `.trim()
   }
