+++ src/app.exercise.js
@@ -1,0 +1,85 @@
+/** @jsx jsx */
+import {jsx} from '@emotion/core'
+
+import React from 'react'
+import * as auth from 'auth-provider'
+// React Router DOM exposes a context provider that all the React Router
+// components use to implicitely access the router data. We need to wrap our
+// AuthenticatedApp in the router
+// 🐨 import the BrowserRouter from 'react-router-dom'
+import {FullPageSpinner} from './components/lib'
+import * as colors from './styles/colors'
+import {client} from './utils/api-client'
+import {useAsync} from './utils/hooks'
+import {AuthenticatedApp} from './authenticated-app'
+import {UnauthenticatedApp} from './unauthenticated-app'
+
+async function getUser() {
+  let user = null
+
+  const token = await auth.getToken()
+  if (token) {
+    const data = await client('me', {token})
+    user = data.user
+  }
+
+  return user
+}
+
+function App() {
+  const {
+    data: user,
+    error,
+    isLoading,
+    isIdle,
+    isError,
+    isSuccess,
+    run,
+    setData,
+  } = useAsync()
+
+  React.useEffect(() => {
+    run(getUser())
+  }, [run])
+
+  const login = form => auth.login(form).then(user => setData(user))
+  const register = form => auth.register(form).then(user => setData(user))
+  const logout = () => {
+    auth.logout()
+    setData(null)
+  }
+
+  if (isLoading || isIdle) {
+    return <FullPageSpinner />
+  }
+
+  if (isError) {
+    return (
+      <div
+        css={{
+          color: colors.danger,
+          height: '100vh',
+          display: 'flex',
+          flexDirection: 'column',
+          justifyContent: 'center',
+          alignItems: 'center',
+        }}
+      >
+        <p>Uh oh... There's a problem. Try refreshing the app.</p>
+        <pre>{error.message}</pre>
+      </div>
+    )
+  }
+
+  if (isSuccess) {
+    const props = {user, login, register, logout}
+    // 🐨 wrap the BrowserRouter around the AuthenticatedApp
+    return user ? (
+      <AuthenticatedApp {...props} />
+    ) : (
+      <UnauthenticatedApp {...props} />
+    )
+  }
+}
+
+export {App}
