--- src/context/index.js
@@ -1,28 +1,0 @@
-import React from 'react'
-import {BrowserRouter as Router} from 'react-router-dom'
-import {ReactQueryConfigProvider} from 'react-query'
-import {AuthProvider} from './auth-context'
-
-const queryConfig = {
-  queries: {
-    useErrorBoundary: true,
-    refetchOnWindowFocus: false,
-    retry(failureCount, error) {
-      if (error.status === 404) return false
-      else if (failureCount < 2) return true
-      else return false
-    },
-  },
-}
-
-function AppProviders({children}) {
-  return (
-    <ReactQueryConfigProvider config={queryConfig}>
-      <Router>
-        <AuthProvider>{children}</AuthProvider>
-      </Router>
-    </ReactQueryConfigProvider>
-  )
-}
-
-export {AppProviders}
