+++ src/index.js
@@ -2,17 +2,8 @@ import {loadDevTools} from './dev-tools/load'
 import './bootstrap'
 import React from 'react'
 import ReactDOM from 'react-dom'
-import {Profiler} from 'components/profiler'
 import {App} from './app'
-import {AppProviders} from './context'
 
 loadDevTools(() => {
-  ReactDOM.render(
-    <Profiler id="App Root" phases={['mount']}>
-      <AppProviders>
-        <App />
-      </AppProviders>
-    </Profiler>,
-    document.getElementById('root'),
-  )
+  ReactDOM.render(<App />, document.getElementById('root'))
 })
