+++ src/screens/not-found.js
@@ -1,23 +1,3 @@
-/** @jsx jsx */
-import {jsx} from '@emotion/core'
+export * from './not-found.final'
 
-import {Link} from 'components/lib'
-
-function NotFoundScreen() {
-  return (
-    <div
-      css={{
-        height: '100%',
-        display: 'grid',
-        alignItems: 'center',
-        justifyContent: 'center',
-      }}
-    >
-      <div>
-        Sorry... nothing here. <Link to="/list">Go home</Link>
-      </div>
-    </div>
-  )
-}
-
-export {NotFoundScreen}
+// export * from './not-found.exercise'
