+++ src/setupProxy.extra-1.js
@@ -1,0 +1,5 @@
+function proxy(app) {
+  app.get(/^\/$/, (req, res) => res.redirect('/discover'))
+}
+
+module.exports = proxy
