+++ src/unauthenticated-app.js
@@ -5,7 +5,6 @@ import React from 'react'
 import {Input, Button, Spinner, FormGroup, ErrorMessage} from './components/lib'
 import {Modal, ModalContents, ModalOpenButton} from './components/modal'
 import {Logo} from './components/logo'
-import {useAuth} from './context/auth-context'
 import {useAsync} from './utils/hooks'
 
 function LoginForm({onSubmit, submitButton}) {
@@ -59,8 +58,7 @@ function LoginForm({onSubmit, submitButton}) {
   )
 }
 
-function UnauthenticatedApp() {
-  const {login, register} = useAuth()
+function UnauthenticatedApp({login, register}) {
   return (
     <div
       css={{
@@ -108,4 +106,4 @@ function UnauthenticatedApp() {
   )
 }
 
-export default UnauthenticatedApp
+export {UnauthenticatedApp}
