+++ src/utils/api-client.js
@@ -1,4 +1,3 @@
-import {queryCache} from 'react-query'
 import * as auth from 'auth-provider'
 const apiURL = process.env.REACT_APP_API_URL
 
@@ -19,7 +18,6 @@ async function client(
 
   return window.fetch(`${apiURL}/${endpoint}`, config).then(async response => {
     if (response.status === 401) {
-      queryCache.clear()
       await auth.logout()
       // refresh the page for them
       window.location.assign(window.location)
